requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
cffi==1.11.5
click==6.7
cryptography==1.9
Flask==1.0.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.8.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.7.2
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
